+++ pyemma/msm/api.py
@@ -53,7 +53,8 @@ __all__ = ['markov_model',
 # MARKOV STATE MODELS - flat Markov chains on discrete observation space
 # =============================================================================
 
-#TODO: show_progress is not documented
+
+# TODO: show_progress is not documented
 @shortcut('its')
 def timescales_msm(dtrajs, lags=None, nits=None, reversible=True, connected=True, weights='empirical',
                    errors=None, nsamples=50, n_jobs=1, show_progress=True):
@@ -133,7 +134,8 @@ def timescales_msm(dtrajs, lags=None, nits=None, reversible=True, connected=True
     ImpliedTimescales
         The object returned by this function.
     pyemma.plots.plot_implied_timescales
-        Implied timescales plotting function. Just call it with the :class:`ImpliedTimescales <pyemma.msm.estimators.ImpliedTimescales>`
+        Implied timescales plotting function. Just call it with the
+        :class:`ImpliedTimescales <pyemma.msm.estimators.ImpliedTimescales>`
         object produced by this function as an argument.
 
 
@@ -184,7 +186,7 @@ def timescales_msm(dtrajs, lags=None, nits=None, reversible=True, connected=True
     else:
         raise ValueError("Weights must be either \'empirical\' or \'oom\'")
     # Set errors to None if weights==oom:
-    if weights=='oom' and (errors is not None):
+    if weights == 'oom' and (errors is not None):
         errors = None
 
     # format data
@@ -197,7 +199,7 @@ def timescales_msm(dtrajs, lags=None, nits=None, reversible=True, connected=True
 
     # Choose estimator:
     if errors is None:
-        if weights=='empirical':
+        if weights == 'empirical':
             estimator = _ML_MSM(reversible=reversible, connectivity=connectivity)
         else:
             estimator = _OOM_MSM(reversible=reversible, connectivity=connectivity)
@@ -300,7 +302,8 @@ def markov_model(P, dt_model='1 step'):
 def estimate_markov_model(dtrajs, lag, reversible=True, statdist=None,
                           count_mode='sliding', weights='empirical',
                           sparse=False, connectivity='largest',
-                          dt_traj='1 step', maxiter=1000000, maxerr=1e-8):
+                          dt_traj='1 step', maxiter=1000000, maxerr=1e-8,
+                          score_method='VAMP2', score_k=10):
     r""" Estimates a Markov model from discrete trajectories
 
     Returns a :class:`MaximumLikelihoodMSM` that
@@ -325,7 +328,7 @@ def estimate_markov_model(dtrajs, lag, reversible=True, statdist=None,
         mode to obtain count matrices from discrete trajectories. Should be
         one of:
 
-        * 'sliding' : A trajectory of length T will have :math:`T-tau` counts
+        * 'sliding' : A trajectory of length T will have :math:`T-\tau` counts
           at time indexes
 
               .. math::
@@ -335,7 +338,7 @@ def estimate_markov_model(dtrajs, lag, reversible=True, statdist=None,
         * 'effective' : Uses an estimate of the transition counts that are
           statistically uncorrelated. Recommended when used with a
           Bayesian MSM.
-        * 'sample' : A trajectory of length T will have :math:`T/tau` counts
+        * 'sample' : A trajectory of length T will have :math:`T/\tau` counts
           at time indexes
 
               .. math::
@@ -376,21 +379,20 @@ def estimate_markov_model(dtrajs, lag, reversible=True, statdist=None,
           states. Estimation will be conducted on the full set of
           states without ensuring connectivity. This only permits
           nonreversible estimation. Currently not implemented.
-    estimate : bool, optional
-        If true estimate the MSM when creating the MSM object.
-    dt : str, optional
+    dt_traj : str, optional
         Description of the physical time corresponding to the lag. May
         be used by analysis algorithms such as plotting tools to
         pretty-print the axes. By default '1 step', i.e. there is no
         physical time unit.  Specify by a number, whitespace and
         unit. Permitted units are (* is an arbitrary string):
 
-        |  'fs',  'femtosecond*'
-        |  'ps',  'picosecond*'
-        |  'ns',  'nanosecond*'
-        |  'us',  'microsecond*'
-        |  'ms',  'millisecond*'
-        |  's',   'second*'
+        *  'fs',  'femtosecond*'
+        *  'ps',  'picosecond*'
+        *  'ns',  'nanosecond*'
+        *  'us',  'microsecond*'
+        *  'ms',  'millisecond*'
+        *  's',   'second*'
+
     maxiter : int, optional
         Optional parameter with reversible = True.  maximum number of
         iterations before the transition matrix estimation method
@@ -405,6 +407,20 @@ def estimate_markov_model(dtrajs, lag, reversible=True, statdist=None,
         order to track changes in small probabilities. The Euclidean
         norm of the change vector, :math:`|e_i|_2`, is compared to
         maxerr.
+
+    score_method : str, optional, default='VAMP2'
+        Score to be used with MSM score function. Available scores are
+        based on the variational approach for Markov processes [13]_ [14]_:
+
+        *  'VAMP1'  Sum of singular values of the symmetrized transition matrix [14]_ .
+                    If the MSM is reversible, this is equal to the sum of transition
+                    matrix eigenvalues, also called Rayleigh quotient [13]_ [15]_ .
+        *  'VAMP2'  Sum of squared singular values of the symmetrized transition matrix [14]_ .
+                    If the MSM is reversible, this is equal to the kinetic variance [16]_ .
+
+    score_k : int or None
+        The maximum number of eigenvalues or singular values used in the
+        score. If set to None, all available eigenvalues will be used.
 
     Returns
     -------
@@ -435,10 +451,10 @@ def estimate_markov_model(dtrajs, lag, reversible=True, statdist=None,
     References
     ----------
     The mathematical theory of Markov (state) model estimation was introduced
-    in [1]_. Further theoretical developments were made in [2]_. The term
-    Markov state model was coined in [3]_. Continuous-time Markov models
+    in [1]_ . Further theoretical developments were made in [2]_ . The term
+    Markov state model was coined in [3]_ . Continuous-time Markov models
     (Master equation models) were suggested in [4]_. Reversible Markov model
-    estimation was introduced in [5]_, and further developed in [6]_,[7]_,[9]_.
+    estimation was introduced in [5]_ , and further developed in [6]_ [7]_ [9]_ .
     It was shown in [8]_ that the quality of Markov state models does in fact
     not depend on memory loss, but rather on where the discretization is
     suitable to approximate the eigenfunctions of the Markov operator (the
@@ -488,6 +504,22 @@ def estimate_markov_model(dtrajs, lag, reversible=True, statdist=None,
     .. [11] Nueske, F., Wu, H., Prinz, J.-H., Wehmeyer, C., Clementi, C. and Noe, F.:
         Markov State Models from short non-Equilibrium Simulations - Analysis and
          Correction of Estimation Bias J. Chem. Phys. (submitted) (2017)
+
+    .. [12] H. Wu and F. Noe: Variational approach for learning Markov processes
+        from time series data (in preparation)
+
+    .. [13] Noe, F. and F. Nueske: A variational approach to modeling slow processes
+        in stochastic dynamical systems. SIAM Multiscale Model. Simul. 11, 635-655 (2013).
+
+    .. [14] Wu, H and F. Noe: Variational approach for learning Markov processes
+        from time series data (in preparation)
+
+    .. [15] McGibbon, R and V. S. Pande: Variational cross-validation of slow
+        dynamical modes in molecular kinetics, J. Chem. Phys. 142, 124105 (2015)
+
+    .. [16] Noe, F. and C. Clementi: Kinetic distance and kinetic maps from molecular
+        dynamics simulation. J. Chem. Theory Comput. 11, 5002-5011 (2015)
+
 
     Example
     -------
@@ -552,7 +584,7 @@ def estimate_markov_model(dtrajs, lag, reversible=True, statdist=None,
                         count_mode=count_mode,
                         sparse=sparse, connectivity=connectivity,
                         dt_traj=dt_traj, maxiter=maxiter,
-                        maxerr=maxerr)
+                        maxerr=maxerr, score_method=score_method, score_k=score_k)
         # estimate and return
         return mlmsm.estimate(dtrajs)
     elif weights == 'oom':
@@ -560,7 +592,8 @@ def estimate_markov_model(dtrajs, lag, reversible=True, statdist=None,
             import warnings
             warnings.warn("Values for statdist, maxiter or maxerr are ignored if OOM-correction is used.")
         oom_msm = _OOM_MSM(lag=lag, reversible=reversible, count_mode=count_mode,
-                           sparse=sparse, connectivity=connectivity, dt_traj=dt_traj)
+                           sparse=sparse, connectivity=connectivity, dt_traj=dt_traj,
+                           score_method=score_method, score_k=score_k)
         # estimate and return
         return oom_msm.estimate(dtrajs)
 
@@ -1083,7 +1116,8 @@ def estimate_hidden_markov_model(dtrajs, nstates, lag, reversible=True, stationa
 
     """
     # initialize HMSM estimator
-    hmsm_estimator = _ML_HMSM(lag=lag, nstates=nstates, reversible=reversible, msm_init='largest-strong',
+    hmsm_estimator = _ML_HMSM(lag=lag, nstates=nstates, reversible=reversible, stationary=stationary,
+                              msm_init='largest-strong',
                               connectivity=connectivity, mincount_connectivity=mincount_connectivity, separate=separate,
                               observe_nonempty=observe_nonempty, stride=stride, dt_traj=dt_traj,
                               accuracy=accuracy, maxit=maxit)
@@ -1248,6 +1282,7 @@ def bayesian_hidden_markov_model(dtrajs, nstates, lag, nsamples=100, reversible=
 
     """
     bhmsm_estimator = _Bayes_HMSM(lag=lag, nstates=nstates, stride=stride, nsamples=nsamples, reversible=reversible,
+                                  stationary=stationary,
                                   connectivity=connectivity, mincount_connectivity=mincount_connectivity,
                                   separate=separate, observe_nonempty=observe_nonempty,
                                   dt_traj=dt_traj, conf=conf, store_hidden=store_hidden, show_progress=show_progress)
@@ -1382,7 +1417,6 @@ def tpt(msmobj, A, B):
         raise ValueError('set A or B defines more states, than given transition matrix.')
 
     # forward committor
-    #msmobj.
     qplus = msmana.committor(T, A, B, forward=True)
     # backward committor
     if msmana.is_reversible(T, mu=mu):