+++ pyemma/msm/tests/test_msm.py
@@ -36,7 +36,7 @@ from msmtools.estimation import count_matrix, largest_connected_set, largest_con
 from msmtools.analysis import stationary_distribution, timescales
 from pyemma.util.numeric import assert_allclose
 from pyemma.msm.tests.birth_death_chain import BirthDeathChain
-from pyemma.msm import estimate_markov_model
+from pyemma.msm import estimate_markov_model, MaximumLikelihoodMSM
 from six.moves import range
 
 
@@ -127,9 +127,7 @@ class TestMSMRevPi(unittest.TestCase):
         self.assertTrue(np.all(msm.active_set==np.array([0, 2])))
         with self.assertRaises(ValueError):
             msm = estimate_markov_model(dtraj_invalid, 1, statdist=pi)
-        
-        
-        
+
 
 class TestMSMDoubleWell(unittest.TestCase):
 
@@ -152,6 +150,43 @@ class TestMSMDoubleWell(unittest.TestCase):
                                                     statdist=cls.statdist,
                                                     sparse=True)
         cls.msm_sparse = estimate_markov_model(cls.dtraj, cls.tau, reversible=False, sparse=True)
+
+    # ---------------------------------
+    # SCORE
+    # ---------------------------------
+
+    def _score(self, msm):
+        dtrajs_test = self.dtraj[80000:]
+        s1 = msm.score(dtrajs_test, score_method='VAMP1', score_k=2)
+        assert 1.0 <= s1 <= 2.0
+        s2 = msm.score(dtrajs_test, score_method='VAMP2', score_k=2)
+        assert 1.0 <= s2 <= 2.0
+        # se = msm.score(dtrajs_test, score_method='VAMPE', score_k=2)
+        # se_inf = msm.score(dtrajs_test, score_method='VAMPE', score_k=None)
+
+    def test_score(self):
+        self._score(self.msmrev)
+        self._score(self.msmrevpi)
+        self._score(self.msm)
+        self._score(self.msmrev_sparse)
+        self._score(self.msmrevpi_sparse)
+        self._score(self.msm_sparse)
+
+    def _score_cv(self, estimator):
+        s1 = estimator.score_cv(self.dtraj, n=5, score_method='VAMP1', score_k=2).mean()
+        assert 1.0 <= s1 <= 2.0
+        s2 = estimator.score_cv(self.dtraj, n=5, score_method='VAMP2', score_k=2).mean()
+        assert 1.0 <= s2 <= 2.0
+        se = estimator.score_cv(self.dtraj, n=5, score_method='VAMPE', score_k=2).mean()
+        se_inf = estimator.score_cv(self.dtraj, n=5, score_method='VAMPE', score_k=None).mean()
+
+    def test_score_cv(self):
+        self._score_cv(MaximumLikelihoodMSM(lag=10, reversible=True))
+        self._score_cv(MaximumLikelihoodMSM(lag=10, reversible=True, statdist_constraint=self.statdist))
+        self._score_cv(MaximumLikelihoodMSM(lag=10, reversible=False))
+        self._score_cv(MaximumLikelihoodMSM(lag=10, reversible=True, sparse=True))
+        self._score_cv(MaximumLikelihoodMSM(lag=10, reversible=True, statdist_constraint=self.statdist, sparse=True))
+        self._score_cv(MaximumLikelihoodMSM(lag=10, reversible=False, sparse=True))
 
     # ---------------------------------
     # BASIC PROPERTIES
@@ -331,7 +366,7 @@ class TestMSMDoubleWell(unittest.TestCase):
         assert (msmana.is_connected(P))
         # REVERSIBLE
         if msm.is_reversible:
-            assert (msmana.is_reversible(P))    
+            assert (msmana.is_reversible(P))
 
     def test_transition_matrix(self):
         self._transition_matrix(self.msmrev)