+++ pyemma/thermo/tests/test_api.py
@@ -252,35 +252,3 @@ class TestMultiTemperature(unittest.TestCase):
             maxiter=10000, maxerr=1.0E-10, estimator='tram', lag=10, connectivity='reversible_pathways')
         validate_thermodynamics(self, tram)
         validate_kinetics(self, tram)
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-