+++ pyemma/util/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import absolute_import
+from ._config import Config as _Config
+
+# default config instance
+config = _Config()